refactor(cli): move completion and setup-env out of core/cli.py
Last migration phase of CLI_REFACTOR_PLAN.md: the `completion` group and
the `setup-env` command move to modules/utils/, leaving core/cli.py as the
assembler it was meant to be (465 -> 152 lines).

- modules/utils/completion.py: `completion` group (238 lines, verbatim).
- modules/utils/system_cli.py: `setup-env` (58 lines, verbatim).

No subpackage was needed here (unlike protocols/): modules/utils/__init__.py
is empty and core/cli.py already imported from ..utils, so the static import
costs nothing at startup.

core/cli.py drops the imports left unused: subprocess, Path and every output
helper but console/STYLES, the only two print_header uses.

test_cli_structure.py extends the 2.3 invariant to the two new files -- they
live straight in utils/ and are not named cli.py, so neither rglob("cli.py")
nor the protocols/cli glob would have covered them.

Verified: zero diff against tests/snapshots/cli_tree_linux.txt; pytest 328
passed, 3 failed (the same TestRunBridge/TestRunSxBridge failures that
predate the refactor).
